remove test module, avoid import, remove whitespace
