+++ mybatis-spring-boot-samples/mybatis-spring-boot-sample-annotation/pom.xml
@@ -42,6 +42,10 @@
 			<artifactId>spring-boot-starter-test</artifactId>
 			<scope>test</scope>
 		</dependency>
+		<dependency>
+			<groupId>org.mybatis.spring.boot</groupId>
+			<artifactId>mybatis-spring-boot-starter-test</artifactId>
+		</dependency>
 	</dependencies>
 	<build>
 		<plugins>
