--- mybatis-spring-boot-samples/mybatis-spring-boot-sample-test/src/main/java/sample/mybatis/SampleTestApplication.java
@@ -1,33 +1,0 @@
-/**
- * Copyright 2015-2017 the original author or authors.
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package sample.mybatis;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- *
- * @author wonwoo
- * @since 1.2.1
- */
-@SpringBootApplication
-public class SampleTestApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(SampleTestApplication.class, args);
-  }
-}
