--- mybatis-spring-boot-samples/mybatis-spring-boot-sample-test/src/main/java/sample/mybatis/domain/City.java
@@ -1,75 +1,0 @@
-/**
- * Copyright 2015-2017 the original author or authors.
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package sample.mybatis.domain;
-
-import java.io.Serializable;
-
-/**
- *
- * @author wonwoo
- * @since 1.2.1
- */
-public class City implements Serializable {
-
-  private static final long serialVersionUID = 1L;
-
-  private Long id;
-
-  private String name;
-
-  private String state;
-
-  private String country;
-
-  public Long getId() {
-    return this.id;
-  }
-
-  public void setId(Long id) {
-    this.id = id;
-  }
-
-  public String getName() {
-    return this.name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  public String getState() {
-    return this.state;
-  }
-
-  public void setState(String state) {
-    this.state = state;
-  }
-
-  public String getCountry() {
-    return this.country;
-  }
-
-  public void setCountry(String country) {
-    this.country = country;
-  }
-
-  @Override
-  public String toString() {
-    return getId() + "," + getName() + "," + getState() + "," + getCountry();
-  }
-
-}
