--- mybatis-spring-boot-samples/mybatis-spring-boot-sample-test/src/main/resources/import.sql
@@ -1,19 +1,0 @@
---        Copyright 2015-2017 the original author or authors.
---
---        Licensed under the Apache License, Version 2.0 (the "License");
---        you may not use this file except in compliance with the License.
---        You may obtain a copy of the License at
---
---           http://www.apache.org/licenses/LICENSE-2.0
---
---        Unless required by applicable law or agreed to in writing, software
---        distributed under the License is distributed on an "AS IS" BASIS,
---        WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
---        See the License for the specific language governing permissions and
---        limitations under the License.
-
-drop table if exists city;
-
-create table city (id int primary key auto_increment, name varchar, state varchar, country varchar);
-
-insert into city (name, state, country) values ('Seoul', 'Seoul KR', 'KOREA');
