--- mybatis-spring-boot-samples/mybatis-spring-boot-sample-test/src/main/resources/sample/mybatis/mapper/CityMapper.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!--
-
-       Copyright 2015-2017 the original author or authors.
-
-       Licensed under the Apache License, Version 2.0 (the "License");
-       you may not use this file except in compliance with the License.
-       You may obtain a copy of the License at
-
-          http://www.apache.org/licenses/LICENSE-2.0
-
-       Unless required by applicable law or agreed to in writing, software
-       distributed under the License is distributed on an "AS IS" BASIS,
-       WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-       See the License for the specific language governing permissions and
-       limitations under the License.
-
--->
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="sample.mybatis.mapper.CityMapper">
-
-    <select id="findByname" resultType="City">
-        select * from city where name = #{name}
-    </select>
-
-    <select id="findById" resultType="City">
-        select * from city where id = #{id}
-    </select>
-
-    <insert id="createCity">
-        insert into city (name, state, country) values(#{name}, #{state}, #{country})
-    </insert>
-
-</mapper>
