+++ mybatis-spring-boot-samples/mybatis-spring-boot-sample-xml/src/test/java/sample/mybatis/SampleSqlSessionTests.java
@@ -1,0 +1,63 @@
+/**
+ *    Copyright 2015-2017 the original author or authors.
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package sample.mybatis;
+
+import org.apache.ibatis.session.SqlSession;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mybatis.spring.boot.test.autoconfigure.MybatisTest;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.test.context.junit4.SpringRunner;
+import sample.mybatis.dao.CityDao;
+import sample.mybatis.domain.City;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+/**
+ * @author wonwoo
+ * @since 1.2.1
+ */
+@RunWith(SpringRunner.class)
+@MybatisTest
+public class SampleSqlSessionTests {
+
+  @Autowired
+  private SqlSession sqlSession;
+
+  @Test
+  public void sqlSessionIsNotNullTest() {
+    assertThat(sqlSession).isNotNull();
+  }
+
+  @Test
+  public void selectCityByIdTest() {
+    City city = sqlSession.selectOne("selectCityById", 1);
+    assertThat(city.getName()).isEqualTo("San Francisco");
+    assertThat(city.getState()).isEqualTo("CA");
+    assertThat(city.getCountry()).isEqualTo("US");
+  }
+
+  @TestConfiguration
+  public static class Config {
+    @Bean
+    public CityDao cityDao(SqlSession sqlSession) {
+      return new CityDao(sqlSession);
+    }
+  }
+}
