+++ mybatis-spring-boot-samples/pom.xml
@@ -29,6 +29,5 @@
   <modules>
     <module>mybatis-spring-boot-sample-annotation</module>
     <module>mybatis-spring-boot-sample-xml</module>
-    <module>mybatis-spring-boot-sample-test</module>
   </modules>
 </project>
