+++ mybatis-spring-boot-test-autoconfigure/src/main/java/org/mybatis/spring/boot/test/autoconfigure/AutoConfigureMybatis.java
@@ -16,9 +16,14 @@
 
 package org.mybatis.spring.boot.test.autoconfigure;
 
-import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Inherited;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
-import java.lang.annotation.*;
+import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
 
 /**
  *
