+++ mybatis-spring-boot-test-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,7 +1,7 @@
 # AutoConfigureDataMybatis auto-configuration imports
 org.mybatis.spring.boot.test.autoconfigure.AutoConfigureMybatis=\
-  org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration,\
-  org.springframework.boot.autoconfigure.jdbc.DataSourceTransactionManagerAutoConfiguration,\
-  org.springframework.boot.autoconfigure.jdbc.JdbcTemplateAutoConfiguration,\
-  org.springframework.boot.autoconfigure.transaction.TransactionAutoConfiguration,\
-  org.mybatis.spring.boot.autoconfigure.MybatisAutoConfiguration
+org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration,\
+org.springframework.boot.autoconfigure.jdbc.DataSourceTransactionManagerAutoConfiguration,\
+org.springframework.boot.autoconfigure.jdbc.JdbcTemplateAutoConfiguration,\
+org.springframework.boot.autoconfigure.transaction.TransactionAutoConfiguration,\
+org.mybatis.spring.boot.autoconfigure.MybatisAutoConfiguration
